+++ GRSlib/motion/scoring.py
@@ -4,6 +4,7 @@ import lammps, lammps.mliap
 from lammps.mliap.loader import *
 from jax import grad, jit
 from functools import partial
+import numpy as np
 
 #Scoring has to be a class within motion because we want a consistent reference for scores, ans this
 #refrence will be LAMMPS using a constructed potential energy surface from the representation loss function
@@ -13,16 +14,19 @@ class Scoring:
     def __init__(self, data, current_desc, target_desc, pt, config):
         self.pt = pt #ParallelTools()
         self.config = config #Config()
-        #Bring in the target and current descriptors here, will be with self. then
-        #descriptors_flt = descriptors.flatten()
         self.current_desc = current_desc
         self.target_desc = target_desc
         self.data = data
-        #self.loss_ff = 0.0 #set to a constant term to initialise
+        self.n_elements = self.config.sections['BASIS'].numtypes
+        if self.n_elements > 1:
+            current_desc = current_desc.flatten()
+            target_desc = target_desc.flatten()
+#        print(np.shape(current_desc),np.shape(target_desc))
 
     @partial(jit, static_argnums=(0))
     def loss_function(self):
         #construct the Jax graph to hand off to lammps
+        #TODO This should be generalized to allow for any loss function defined in lossfunc/ to be called
         first_mom = None
         second_mom = None
         third_mom = None
@@ -46,9 +50,7 @@ class Scoring:
         loss_ff = first_mom
         for item in set_of_moments:
              loss_ff += item
-#        return loss_ff, grad(loss_ff)
         return loss_ff
-    #, grad(first_mom)
 
     def construct_lmp(self):
         #Generates the major components of a lammps script needed for a scoring call
@@ -65,16 +67,17 @@ class Scoring:
         self._lmp.command("atom_style atomic")
         self._lmp.command("read_data %s" % self.data)
         #TODO make the possibility to import any reference potential to be used with the mliap one
-#            self._lmp.command("pair_style hybrid/overlay soft %2.3f mliap model mliappy LATER descriptor ace coupling_coefficients.yace" % something)
-#            self._lmp.command("pair_coeff * * soft %f" % something)
-        self._lmp.command("pair_style mliap model mliappy LATER descriptor ace coupling_coefficients.yace")
+#        self._lmp.command("pair_style mliap model mliappy LATER descriptor ace coupling_coefficients.yace")
 
+        self._lmp.command("pair_style hybrid/overlay soft %2.3f mliap model mliappy LATER descriptor ace coupling_coefficients.yace" % self.config.sections['MOTION'].soft_strength)
+        self._lmp.command("pair_coeff * * soft %f" % self.config.sections['MOTION'].soft_strength)
+        self._lmp.command("pair_coeff mliap * * %s" % (" ".join(str(x) for x in self.config.sections['BASIS'].elements)))
         self._lmp.command('neighbor  2.3 bin')
         self._lmp.command("neigh_modify one 10000")
-#        loss_ff, grad_loss_ff = self.loss_function()
         loss_ff = self.loss_function()
+        forces = grad(self.loss_function)
         lammps.mliap.load_model(loss_ff)
-
+              
     def get_atomic_energies(self):
         #Return as array per-atom energies for the set of potentials applied
         self.construct_lmp()
@@ -95,7 +98,7 @@ class Scoring:
         del self._lmp
         return atom_forces
 
-    def get_score(self):
+    def get_score(self,add_cmds):
         #Return as array unweighted scores per moment
         self.construct_lmp()
         self._lmp.commands_string("run 0")
@@ -103,8 +106,18 @@ class Scoring:
         del self._lmp
         return score
 
+    def add_cmds_before_score(self,string):
+        self.construct_lmp()
+        self._lmp.commands_string("run 0")
+        before_score = self._lmp.get_thermo("pe") # potential energy
+        self._extract_commands(string)
+        self._lmp.commands_string("run 0")
+        after_score = self._lmp.get_thermo("pe") # potential energy
+        del self._lmp
+        return before_score, after_score
+
     def _extract_commands(self,string):
         #Can be given a block of text where it will split them into individual commands
-        lmp_setup = [x for x in string.splitlines() if x.strip() != '']
-        for line in lmp_setup:
+        add_lmp_lines = [x for x in string.splitlines() if x.strip() != '']
+        for line in add_lmp_lines:
                 self._lmp.command(line)
